We didn't need this as

+++ src/compact_representation/core/cell_board/eval.rs
@@ -140,7 +140,7 @@ impl<T: CellNum, D: Dimensions, const BOARD_SIZE: usize, const MAX_SNAKES: usize
                 new_health = new_health.saturating_sub(1);
                 if self.get_cell(new_head).is_hazard() {
                     new_health = if self.hazard_damage.is_negative() {
-                        new_health.saturating_add(self.hazard_damage.unsigned_abs() as u8)
+                        new_health.saturating_add(self.hazard_damage.unsigned_abs())
                     } else {
                         new_health.saturating_sub(self.hazard_damage as u8)
                     };
